fish-kms/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
cryptography==42.0.0
opencv-python==4.9.0.80
python-dotenv==1.0.0
numpy==1.26.3

